fix: Swiping cards on mobile was very laggy, I hope this update improves that!
I improved the performance by using translate3d instead of using top and left to change the card position. So now the webapp will utilize the device GPU instead of CPU. Also, changing the left and top position caused a reflow every time they were updated, which causes a huge performance drain. But using translate3d remedies that, because it does not trigger a full on recalculation of the page layout.